+++ apps/backend/src/main/java/io/github/artsobol/kurkod/model/request/breed/BreedPostRequest.java
@@ -16,8 +16,8 @@ public class BreedPostRequest implements Serializable {
     private String name;
 
     @NotNull
-    private short eggsNumber;
+    private Integer eggsNumber;
 
     @NotNull
-    private short weight;
+    private Integer weight;
 }
