+++ apps/backend/src/main/java/io/github/artsobol/kurkod/model/request/chicken/ChickenPutRequest.java
@@ -13,7 +13,7 @@ import java.time.LocalDate;
 public class ChickenPutRequest {
 
     @NotNull
-    private short weight;
+    private Integer weight;
 
     @NotNull
     private LocalDate birthDate;
